refactor: postList 관련 api 를 postList.ts 로 옮김
