+++ src/app/page.tsx
@@ -5,7 +5,7 @@ import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import { useSearchParams } from 'next/navigation';
 
-import { PostList, getPostList } from '@/src/apis/post';
+import { PostList, getPostList } from '@/src/apis/postList';
 import {
   PostContentToPostItemType,
   searchParamsToGetPostListParams,
